requirements.txt
torch>=1.13.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
datasets>=2.7.0
accelerate>=0.18.0
peft>=0.4.0
evaluate>=0.4.3
rouge_score
sacrebleu
bert_score